+++ app/lib/features/squads/presentation/state/squad_settings_notifier.dart
@@ -96,7 +96,7 @@ class SquadSettingsNotifier extends AsyncNotifier<List<SquadMember>> {
 
   Future<void> updateRankingSettings({
     required bool rankingUpdate,
-    required double rankingMultiplier,
+    required int rankingMultiplier,
     required bool useExperienceFactor,
   }) async {
     state = const AsyncValue.loading();
